# [DATA-661] Stop attempting to paint new maps when optimizing the map