--- a/slam-libraries/viam-cartographer/src/io/submap_painter.h
+++ b/slam-libraries/viam-cartographer/src/io/submap_painter.h
@@ -15,14 +15,22 @@
 namespace viam {
 namespace io {
 
+// PaintSubmapSlices paints the submaps using cairo and returns
+// the result in form of PaintSubmapsSlicesResult.
 cartographer::io::PaintSubmapSlicesResult PaintSubmapSlices(
     const std::map<::cartographer::mapping::SubmapId,
                    cartographer::io::SubmapSlice>& submaps,
     double resolution);
 
+// DrawPoseOnSurface draws global_pose onto the painted_slices.
 void DrawPoseOnSurface(
     cartographer::io::PaintSubmapSlicesResult* painted_slices,
     cartographer::transform::Rigid3d global_pose, float resolution);
+
+// WritePaintedSlicesToJpegString writes painted slices to
+// a jpeg string and returns it.
+std::string WritePaintedSlicesToJpegString(
+    cartographer::io::PaintSubmapSlicesResult& painted_slices);
 
 }  // namespace io
 }  // namespace viam