--- a/slam-libraries/viam-cartographer/src/slam_service/slam_service.cc
+++ b/slam-libraries/viam-cartographer/src/slam_service/slam_service.cc
@@ -211,7 +211,8 @@ void SLAMServiceImpl::PaintMap(bool pose_marker_flag) {
         submap_slices;
 
     if (submap_poses.size() == 0) {
-        return "";
+        jpeg_img = "";
+        return;
     }
 
     for (const auto &&submap_id_pose : submap_poses) {