--- a/slam-libraries/viam-cartographer/src/slam_service/slam_service.cc
+++ b/slam-libraries/viam-cartographer/src/slam_service/slam_service.cc
@@ -24,7 +24,7 @@ std::atomic<bool> b_continue_session{true};
                                             GetPositionResponse *response) {
     cartographer::transform::Rigid3d global_pose;
     {
-        std::lock_guard<std::mutex> lk(viam_response_mutex);
+        std::lock_guard<std::mutex> lk(latest_global_pose_mutex);
         global_pose = latest_global_pose;
     }
     // Setup mapping of pose message to the response. NOTE not using
@@ -85,16 +85,19 @@ std::atomic<bool> b_continue_session{true};
         if (!optimizing) {
             UpdateLatestSubmapSlices();
         }
-        if (latest_submap_slices.size() == 0) {
-            return grpc::Status(grpc::StatusCode::UNAVAILABLE,
-                                "currently no map exists yet");
+        {
+            std::lock_guard<std::mutex> lk(latest_submap_slices_mutex);
+            if (latest_submap_slices.size() == 0) {
+                return grpc::Status(grpc::StatusCode::UNAVAILABLE,
+                                    "currently no map exists yet");
+            }
+            cartographer::io::PaintSubmapSlicesResult painted_slices =
+                viam::io::PaintSubmapSlices(latest_submap_slices, kPixelSize);
+            if (pose_marker_flag) {
+                PaintMarker(painted_slices);
+            }
+            jpeg_img = viam::io::WritePaintedSlicesToJpegString(painted_slices);
         }
-        cartographer::io::PaintSubmapSlicesResult painted_slices =
-            viam::io::PaintSubmapSlices(latest_submap_slices, kPixelSize);
-        if (pose_marker_flag) {
-            PaintMarker(painted_slices);
-        }
-        jpeg_img = viam::io::WritePaintedSlicesToJpegString(painted_slices);
     } catch (std::exception &e) {
         std::ostringstream oss;
         oss << "error encoding image " << e.what();
@@ -114,15 +117,10 @@ std::atomic<bool> b_continue_session{true};
 
 ::grpc::Status SLAMServiceImpl::GetPointCloudMap(const GetMapRequest *request,
                                                  GetMapResponse *response) {
-    // Write the pointcloud to the pointcloud_buffer
+    // Write the pointcloud to the latest_pointcloud_buffer
     try {
         if (!optimizing) {
             ExtractPointCloudToBuffer();
-        }
-        if (!pointcloud_has_points) {
-            LOG(ERROR) << "map pointcloud does not have points yet";
-            return grpc::Status(grpc::StatusCode::UNAVAILABLE,
-                                "map pointcloud does not have points yet");
         }
     } catch (std::exception &e) {
         std::ostringstream oss;
@@ -133,7 +131,15 @@ std::atomic<bool> b_continue_session{true};
     // Write the pointcloud to the response
     try {
         common::v1::PointCloudObject *pco = response->mutable_point_cloud();
-        pco->set_point_cloud(pointcloud_buffer.str());
+        {
+            std::lock_guard<std::mutex> lk(latest_pointcloud_mutex);
+            if (!latest_pointcloud_has_points) {
+                LOG(ERROR) << "map pointcloud does not have points yet";
+                return grpc::Status(grpc::StatusCode::UNAVAILABLE,
+                                    "map pointcloud does not have points yet");
+            }
+            pco->set_point_cloud(latest_pointcloud_buffer.str());
+        }
         return grpc::Status::OK;
     } catch (std::exception &e) {
         std::ostringstream oss;
@@ -216,48 +222,51 @@ void SLAMServiceImpl::UpdateLatestSubmapSlices() {
         }
     }
 
-    latest_submap_slices = std::map<cartographer::mapping::SubmapId,
-                                    ::cartographer::io::SubmapSlice>{};
+    {
+        std::lock_guard<std::mutex> lk(latest_submap_slices_mutex);
+        latest_submap_slices = std::map<cartographer::mapping::SubmapId,
+                                        ::cartographer::io::SubmapSlice>{};
 
-    if (submap_poses.size() == 0) {
-        return;
-    }
-
-    for (const auto &&submap_id_pose : submap_poses) {
-        auto submap_textures =
-            absl::make_unique<::cartographer::io::SubmapTextures>();
-        submap_textures->version =
-            response_protos[submap_id_pose.id].submap_version();
-        for (const auto &texture_proto :
-             response_protos[submap_id_pose.id].textures()) {
-            const std::string compressed_cells(texture_proto.cells().begin(),
-                                               texture_proto.cells().end());
-            submap_textures->textures.emplace_back(
-                ::cartographer::io::SubmapTexture{
-                    ::cartographer::io::UnpackTextureData(
-                        compressed_cells, texture_proto.width(),
-                        texture_proto.height()),
-                    texture_proto.width(), texture_proto.height(),
-                    texture_proto.resolution(),
-                    cartographer::transform::ToRigid3(
-                        texture_proto.slice_pose())});
+        if (submap_poses.size() == 0) {
+            return;
         }
 
-        // Prepares SubmapSlice
-        ::cartographer::io::SubmapSlice &submap_slice =
-            latest_submap_slices[submap_id_pose.id];
-        const auto fetched_texture = submap_textures->textures.begin();
-        submap_slice.pose = submap_id_pose.data.pose;
-        submap_slice.width = fetched_texture->width;
-        submap_slice.height = fetched_texture->height;
-        submap_slice.slice_pose = fetched_texture->slice_pose;
-        submap_slice.resolution = fetched_texture->resolution;
-        submap_slice.cairo_data.clear();
+        for (const auto &&submap_id_pose : submap_poses) {
+            auto submap_textures =
+                absl::make_unique<::cartographer::io::SubmapTextures>();
+            submap_textures->version =
+                response_protos[submap_id_pose.id].submap_version();
+            for (const auto &texture_proto :
+                response_protos[submap_id_pose.id].textures()) {
+                const std::string compressed_cells(texture_proto.cells().begin(),
+                                                texture_proto.cells().end());
+                submap_textures->textures.emplace_back(
+                    ::cartographer::io::SubmapTexture{
+                        ::cartographer::io::UnpackTextureData(
+                            compressed_cells, texture_proto.width(),
+                            texture_proto.height()),
+                        texture_proto.width(), texture_proto.height(),
+                        texture_proto.resolution(),
+                        cartographer::transform::ToRigid3(
+                            texture_proto.slice_pose())});
+            }
 
-        submap_slice.surface = ::cartographer::io::DrawTexture(
-            fetched_texture->pixels.intensity, fetched_texture->pixels.alpha,
-            fetched_texture->width, fetched_texture->height,
-            &submap_slice.cairo_data);
+            // Prepares SubmapSlice
+            ::cartographer::io::SubmapSlice &submap_slice =
+                latest_submap_slices[submap_id_pose.id];
+            const auto fetched_texture = submap_textures->textures.begin();
+            submap_slice.pose = submap_id_pose.data.pose;
+            submap_slice.width = fetched_texture->width;
+            submap_slice.height = fetched_texture->height;
+            submap_slice.slice_pose = fetched_texture->slice_pose;
+            submap_slice.resolution = fetched_texture->resolution;
+            submap_slice.cairo_data.clear();
+
+            submap_slice.surface = ::cartographer::io::DrawTexture(
+                fetched_texture->pixels.intensity, fetched_texture->pixels.alpha,
+                fetched_texture->width, fetched_texture->height,
+                &submap_slice.cairo_data);
+        }
     }
 }
 
@@ -265,17 +274,13 @@ void SLAMServiceImpl::PaintMarker(
     cartographer::io::PaintSubmapSlicesResult &painted_slices) {
     cartographer::transform::Rigid3d global_pose;
     {
-        std::lock_guard<std::mutex> lk(viam_response_mutex);
+        std::lock_guard<std::mutex> lk(latest_global_pose_mutex);
         global_pose = latest_global_pose;
     }
     viam::io::DrawPoseOnSurface(&painted_slices, global_pose, kPixelSize);
 }
 
 void SLAMServiceImpl::ExtractPointCloudToBuffer() {
-    pointcloud_buffer = std::stringbuf();
-    std::ostream oss(&pointcloud_buffer);
-    pointcloud_has_points = false;
-
     cartographer::mapping::MapById<cartographer::mapping::NodeId,
                                    cartographer::mapping::TrajectoryNode>
         trajectory_nodes;
@@ -285,51 +290,58 @@ void SLAMServiceImpl::ExtractPointCloudToBuffer() {
             map_builder.map_builder_->pose_graph()->GetTrajectoryNodes();
     }
 
-    long number_points = 0;
-    for (const auto &&trajectory_node : trajectory_nodes) {
-        auto point_cloud = trajectory_node.data.constant_data
-                               ->filtered_gravity_aligned_point_cloud;
-        number_points += point_cloud.size();
-    }
+    {
+        std::lock_guard<std::mutex> lk(latest_pointcloud_mutex);
+        latest_pointcloud_buffer = std::stringbuf();
+        std::ostream oss(&latest_pointcloud_buffer);
+        latest_pointcloud_has_points = false;
 
-    // Write our PCD file, which is written as a binary.
-    oss << "VERSION .7\n"
-        << "FIELDS x y z rgb\n"
-        << "SIZE 4 4 4 4\n"
-        << "TYPE F F F I\n"
-        << "COUNT 1 1 1 1\n"
-        << "WIDTH " << number_points << "\n"
-        << "HEIGHT " << 1 << "\n"
-        << "VIEWPOINT 0 0 0 1 0 0 0\n"
-        << "POINTS " << number_points << "\n"
-        << "DATA binary\n";
+        long number_points = 0;
+        for (const auto &&trajectory_node : trajectory_nodes) {
+            auto point_cloud = trajectory_node.data.constant_data
+                                ->filtered_gravity_aligned_point_cloud;
+            number_points += point_cloud.size();
+        }
 
-    for (const auto &&trajectory_node : trajectory_nodes) {
-        cartographer::sensor::PointCloud local_gravity_aligned_point_cloud =
-            trajectory_node.data.constant_data
-                ->filtered_gravity_aligned_point_cloud;
+        // Write our PCD file, which is written as a binary.
+        oss << "VERSION .7\n"
+            << "FIELDS x y z rgb\n"
+            << "SIZE 4 4 4 4\n"
+            << "TYPE F F F I\n"
+            << "COUNT 1 1 1 1\n"
+            << "WIDTH " << number_points << "\n"
+            << "HEIGHT " << 1 << "\n"
+            << "VIEWPOINT 0 0 0 1 0 0 0\n"
+            << "POINTS " << number_points << "\n"
+            << "DATA binary\n";
 
-        // We're only applying the translation of the `global_pose` on the
-        // point cloud here, as opposed to using both the translation and
-        // rotation of the global pose of the trajectory node. The reason for
-        // this seems to be that since the point cloud is already gravity
-        // aligned, the rotational part of the transformation relative to the
-        // world frame is already taken care of.
-        cartographer::sensor::PointCloud global_point_cloud =
-            cartographer::sensor::TransformPointCloud(
-                local_gravity_aligned_point_cloud,
-                cartographer::transform::Rigid3f(
-                    trajectory_node.data.global_pose.cast<float>()
-                        .translation(),
-                    cartographer::transform::Rigid3f::Quaternion::Identity()));
+        for (const auto &&trajectory_node : trajectory_nodes) {
+            cartographer::sensor::PointCloud local_gravity_aligned_point_cloud =
+                trajectory_node.data.constant_data
+                    ->filtered_gravity_aligned_point_cloud;
 
-        for (auto &&point : global_point_cloud) {
-            int rgb = 0;
-            pointcloud_buffer.sputn((const char *)&point.position[1], 4);
-            pointcloud_buffer.sputn((const char *)&point.position[2], 4);
-            pointcloud_buffer.sputn((const char *)&point.position[0], 4);
-            pointcloud_buffer.sputn((const char *)&rgb, 4);
-            pointcloud_has_points = true;
+            // We're only applying the translation of the `global_pose` on the
+            // point cloud here, as opposed to using both the translation and
+            // rotation of the global pose of the trajectory node. The reason for
+            // this seems to be that since the point cloud is already gravity
+            // aligned, the rotational part of the transformation relative to the
+            // world frame is already taken care of.
+            cartographer::sensor::PointCloud global_point_cloud =
+                cartographer::sensor::TransformPointCloud(
+                    local_gravity_aligned_point_cloud,
+                    cartographer::transform::Rigid3f(
+                        trajectory_node.data.global_pose.cast<float>()
+                            .translation(),
+                        cartographer::transform::Rigid3f::Quaternion::Identity()));
+
+            for (auto &&point : global_point_cloud) {
+                int rgb = 0;
+                latest_pointcloud_buffer.sputn((const char *)&point.position[1], 4);
+                latest_pointcloud_buffer.sputn((const char *)&point.position[2], 4);
+                latest_pointcloud_buffer.sputn((const char *)&point.position[0], 4);
+                latest_pointcloud_buffer.sputn((const char *)&rgb, 4);
+                latest_pointcloud_has_points = true;
+            }
         }
     }
 }
@@ -542,7 +554,7 @@ void SLAMServiceImpl::ProcessDataAndStartSavingMaps(double data_start_time) {
         }
         // Save a copy of the global pose
         {
-            std::lock_guard<std::mutex> lk(viam_response_mutex);
+            std::lock_guard<std::mutex> lk(latest_global_pose_mutex);
             latest_global_pose = tmp_global_pose;
         }
 
@@ -561,6 +573,8 @@ void SLAMServiceImpl::ProcessDataAndStartSavingMaps(double data_start_time) {
         map_builder.map_builder_->FinishTrajectory(trajectory_id);
     }
     if (offline_flag) {
+        UpdateLatestSubmapSlices();
+        ExtractPointCloudToBuffer();
         {
             optimizing = true;
             std::lock_guard<std::mutex> lk(map_builder_mutex);
@@ -577,7 +591,7 @@ void SLAMServiceImpl::ProcessDataAndStartSavingMaps(double data_start_time) {
         }
 
         {
-            std::lock_guard<std::mutex> lk(viam_response_mutex);
+            std::lock_guard<std::mutex> lk(latest_global_pose_mutex);
             latest_global_pose = tmp_global_pose;
         }
         finished_processing_offline = true;