Ticket: https://viam.atlassian.net/browse/DATA-661

**Problem:**
We optimize the map in two situations:
* At startup, when an apriori map is loaded and `optimize_on_start = true`.
* At the end of processing an offline dataset.

The `map_builder_mutex` is held while the optimization is happening, so grabbing the map in `jpeg` or `pcd` form would have to wait until the optimization is finished. This results in the user having to wait a long time for the map to arrive after requesting it. 

**Solution:**
This solution checks whether or not the optimization is ongoing, and if so, `GetMap` will return the most recently saved map. 

Reasoning: The map is not changing while the optimization is ongoing. As a result, if the user requests a map during the optimization, we can return the most recently created map. 

**Comparing times:**
I tested and timed mapping before and after the changes I've made to ensure that mapping and requesting & receiving maps do not slow down. I used the [viam-old-office-small-pcd](https://storage.googleapis.com/viam-labs-datasets/viam-old-office-small-pcd.zip) dataset.

The times for "before" --> "after" the changes I made are reduced on average:
* Total runtime for mapping the entire dataset: 25 seconds --> 25 seconds
* `slam::server::GetMap`: 784 ms --> 501 ms
* `slam::builtin::GetMap`: 698 ms --> 416 ms